Implement Support for no_std
This allows using rusttype on no_std targets by disabling the `std`
feature that is enabled by default and activating the `libm-math`
feature. On most targets you still want to keep the `has-atomics`
feature active, as that specifies that the target is capable of
supporting atomics, which only some embedded devices do not support.